agent/pyproject.toml
[project]
name = "nexus-agent"
version = "0.1.0"
description = "Agent Planner para o Nexus Event Gateway — gera planos de eventos Exactly-Once"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4",
    "confluent-kafka>=2.6",
    "pydantic>=2.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
